detalle

+++ HomeBanking/Repositories/IAccountRepository.cs
@@ -7,6 +7,6 @@ namespace HomeBanking.Repositories
         IEnumerable<Account> GetAllAccounts();
         void Save(Account account);
         Account FindById(long id);
-        Account FindByIdAndClientEmail(int id, string email);
+        Account FindByIdAndClientEmail(long id, string email);
     }
 }
